feat: mudança nas relaçoes de pedidosalgado e view do salgado funcional

+++ src/main/java/com/example/sisapsoo/model/PedidoSalgado.java
@@ -52,7 +52,7 @@ public class PedidoSalgado {
         this.quantidade = quantidade;
     }
 
-    public float calcularSubtotal() {
+    public double calcularSubtotal() {
         return salgado.getPreco() * quantidade;
     }
 }
